+++ ydb/core/persqueue/public/codecs/ya.make
@@ -7,6 +7,7 @@ SRCS(
 PEERDIR(
     ydb/public/api/protos
     ydb/public/sdk/cpp/src/client/persqueue_public
+    ydb/public/sdk/cpp/src/client/topic
 )
 
 END()
